+++ common/resolve_options.go
@@ -1,8 +1,8 @@
 package common
 
 type ResolveOptions struct {
-	Connection string
-	User       string
-	Session    string
-	Channel    string
+	Connection string `form:"id"`
+	User       string `form:"user"`
+	Session    string `form:"session"`
+	Channel    string `form:"channel"`
 }