+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+- Improve query binding
+
 ## v0.3.0 - 2020-05-08
 
 - **Breaking**